Remove txType validation in GetTransactions
The code attempts to validate transaction types, which will just cause issues going forward.  If we remove validation, and the dev requests a txType that doesn't exist, the code fails just as before.  Adding validation in this wrapper doesn't add anything (except it does hide the actual error with a vague "Validation error" message).
Removing this line fixes two issues:
1. Incorrect "isRequired" validation - txTypes is not a mandatory parameter
2. Incorrect txTypes options subset - there are many more transaction types than Brokerage and Trade.  See here:
https://github.com/independentreserve/dotNetApiClient/blob/b314952ef35ca65758bd826e2b735d33ca227e1f/src/DotNetClientApi/Data/TransactionType.cs#L12